- Restructuring Program 3 --Fixed some bug in '_main_tranform'
